Use CMake instead of Make(+scons)
Palabos fully dropped support for SCons after the release of v2.2.0.
Therefore, I used CMake, as both Palabos and LAMMPS nowadays use CMake
as their build tool.

The `CMakeLists.txt` file is inspired by the `cavity2d` case in Palabos.
It finds all the libraries and links `embolism` against them.

The paths to the libraries need to be specified:
- Here, I hardcoded the paths to branches of bare repositories in the
  `CMakeLists.txt`.
- Alternatively, it is possible to specify the following environment
  variables:
  - `PROJ_PALABOS_ROOT`
  - `PROJ_LAMMPS_ROOT`
  - `PROJ_PALABOS_LAMMPS_ROOT`